QCamera2: Removed unnecessary metadata type casts
- Removed all unnecessary pointer type casts when doing
  translation of metadata structures between camera backend
  and Android native interface.
- Removed obsolete file 'cam_intf.c'.
- Updated structure 'parm_data_t' by adding missing
  fields and corrected some types.
- Using a macro (ADD_SET_PARAM_ENTRY_TO_BATCH) to access
  metadata structure. By adding it, data type compatibility
  will be checked at build time, instead of checking
  at runtime.
- Optimized code by replacing
  'IS_META_AVAILABLE/POINTER_OF_META' sequence with
  a new macro called 'IF_META_AVAILABLE'.
- Removed some obsolete members from HWI classes.
- Depends on change I4f7e7306 and I65b7c1c4.

+++ QCamera2/stack/mm-camera-interface/src/mm_camera_interface.c
@@ -630,7 +630,7 @@ static int32_t mm_camera_intf_link_stream(uint32_t camera_handle,
     uint32_t id = 0;
     mm_camera_obj_t * my_obj = NULL;
 
-    CDBG("%s : E handle = %d ch_id = %d",
+    CDBG("%s : E handle = %u ch_id = %u",
          __func__, camera_handle, ch_id);
 
     pthread_mutex_lock(&g_intf_lock);
@@ -644,8 +644,8 @@ static int32_t mm_camera_intf_link_stream(uint32_t camera_handle,
         pthread_mutex_unlock(&g_intf_lock);
     }
 
-    CDBG("%s :X stream_id = %d", __func__, stream_id);
-    return id;
+    CDBG("%s :X stream_id = %u", __func__, stream_id);
+    return (int32_t)id;
 }
 
 /*===========================================================================
